Rename classes to models
