+++ src/components/hud/TileInfo.js
@@ -1,6 +1,6 @@
 const m = require("mithril");
 
-import Terrain from "../../classes/Terrain.js";
+import Terrain from "../../models/Terrain.js";
 import Button from "../templates/Button.js";
 import UnitCard from "../templates/UnitFlag.js";
