+++ src/models/ColorSolver.js
