+++ src/models/ScenarioDefinition.js
