+++ src/models/ScenarioManifest.js
