+++ src/pages/mainmenu.js
@@ -1,6 +1,6 @@
 const m = require('mithril');
 
-import Page from '../classes/Page.js';
+import Page from '../models/Page.js';
 import Body from "../components/Body.js";
 import Nav from "../components/Nav.js";
 import NavTitle from "../components/templates/NavTitle.js";
