+++ src/pages/scenarios.js
@@ -1,7 +1,7 @@
 const m = require('mithril');
 const classnames = require('classnames');
 
-import Page from '../classes/Page.js';
+import Page from '../models/Page.js';
 import Body from "../components/Body.js";
 import Button from "../components/templates/Button.js";
 import scenarioManifestStore from "../stores/ScenarioManifest.store.js";
