+++ src/utilities/hexToSVGFilter.js
@@ -1,7 +1,7 @@
 // Credit to MultiplyByZer0: Accessed, 2/15/2023
 // (https://stackoverflow.com/questions/42966641/how-to-transform-black-into-any-given-color-using-only-css-filters/43960991#43960991)
-import Color from "../classes/Color.js";
-import ColorSolver from "../classes/ColorSolver.js";
+import Color from "../models/Color.js";
+import ColorSolver from "../models/ColorSolver.js";
 import {hexToRgb} from "./hexToRGB.js";
